Make request to Zillow API GetDeepSearchResults Web Service

The GetDeepSearchResults API finds a property for a specified address. The result set returned contains the full address(s), zpid and Zestimate data that is provided by the GetSearchResults API. Moreover, this API call also gives rich property data like lot size, year built, bath/beds, last sale details etc.

GetDeepSearchResults(address = NULL, citystatezip = NULL,
  rentzestimate = FALSE, zws_id = getOption("ZillowR-zws_id"),
  url = "http://www.zillow.com/webservice/GetDeepSearchResults.htm")

Arguments

address
The address of the property to search. Required.

citystatezip
The city+state combination and/or ZIP code for which to search. Required.

rentzestimate
Return Rent Zestimate information if available (logical, default: false).

zws_id
The Zillow Web Service Identifier. Required.

url
URL for the GetDeepSearchResults Web Service. Required.

Value

A named list with the following elements:

request

a list with the request parameters

message

a list of status code(s) and message(s) returned by the API

response

an XMLNode with the API-specific response values. At this time, no further coercion is performed, so you may have to use functions from the XML package to extract the desired output.

Examples

# NOT RUN {
GetDeepSearchResults(address = '2114 Bigelow Ave', citystatezip = 'Seattle, WA')
GetDeepSearchResults(address = '2114 Bigelow Ave', citystatezip = 'Seattle, WA',
                     rentzestimate = TRUE)
# }