Make request to Zillow API GetZestimate Web Service

For a specified Zillow property identifier (zpid), the GetZestimate API returns:

GetZestimate(zpid = NULL, rentzestimate = FALSE,
  zws_id = getOption("ZillowR-zws_id"),
  url = "http://www.zillow.com/webservice/GetZestimate.htm")

Arguments

zpid
The Zillow Property ID for the property for which to obtain information. Required.

rentzestimate
Return Rent Zestimate information if available (logical, default: false).

zws_id
The Zillow Web Service Identifier. Required.

url
URL for the GetZestimate Web Service. Required.

Value

A named list with the following elements:

request

a list with the request parameters

message

a list of status code(s) and message(s) returned by the API

response

an XMLNode with the API-specific response values. At this time, no further coercion is performed, so you may have to use functions from the XML package to extract the desired output.

Details

The most recent property Zestimate

The date the Zestimate was computed

The valuation range

The Zestimate ranking within the property's ZIP code.

The full property address and geographic location (latitude/longitude) and a set of identifiers that uniquely represent the region (ZIP code, city, county & state) in which the property exists.

The GetZestimate API will only surface properties for which a Zestimate exists. If a request is made for a property that has no Zestimate, an error code is returned. Zillow doesn't have Zestimates for all the homes in its database. For such properties, we do have tax assessment data, but that is not provided through the API. For more information, see our Zestimate coverage.

Examples

# NOT RUN {
GetZestimate(zpid = 48749425)
GetZestimate(zpid = 48749425, rentzestimate = TRUE)
# }